+++ lib/sstls/archive.go
@@ -5,7 +5,7 @@ package sstls
  * Read and Save certs with an archive file
  * By J. Stuart McMurray
  * Created 20240327
- * Last Modified 20260816
+ * Last Modified 20260817
  */
 
 import (
@@ -18,6 +18,13 @@ import (
 	"time"
 
 	"golang.org/x/tools/txtar"
+)
+
+// Names of files in a txtar archive for the PEM-encoded cert and key.
+const (
+	txtarCertFile        = "cert"
+	txtarFingerprintFile = "fingerprint"
+	txtarKeyFile         = "key"
 )
 
 // LoadCachedCertificate loads the certificate from the named file, which
@@ -81,6 +88,29 @@ func loadCachedCertificate(certFile string, b []byte) (tls.Certificate, error) {
 // SaveCertificate saves PEM to the given file.  Directories will be created
 // as needed with 0755 permissions.
 func SaveCertificate(certFile string, certPEM, keyPEM []byte) error {
+	/* Work out the fingerprint.  Also validates the cert for us. */
+	cert, err := tls.X509KeyPair(certPEM, keyPEM)
+	if nil != err {
+		return fmt.Errorf("parsing certificate: %w", err)
+	}
+	fp, err := PubkeyFingerprintTLS(cert)
+	if nil != err {
+		return fmt.Errorf("calculating fingerprint: %w", err)
+	}
+
+	/* Do what we said we'd do in the first place. */
+	return saveCertificateFingerprint(certFile, fp, certPEM, keyPEM)
+}
+
+// saveCertificateFingerprint is like SaveCertificate, but with a
+// pre-calculated fingerprint to avoid having to parse fresh PEM.
+func saveCertificateFingerprint(
+	certFile string,
+	fingerprint string,
+	certPEM []byte,
+	keyPEM []byte,
+) error {
+	/* openFile tries to open certFile. */
 	openFile := func() (*os.File, error) {
 		return os.OpenFile(
 			certFile,
@@ -112,6 +142,9 @@ func SaveCertificate(certFile string, certPEM, keyPEM []byte) error {
 			time.Now().Format(time.RFC3339),
 		),
 		Files: []txtar.File{{
+			Name: txtarFingerprintFile,
+			Data: []byte(fingerprint),
+		}, {
 			Name: txtarCertFile,
 			Data: certPEM,
 		}, {